+++ src/components/DishListComponent.tsx
@@ -1,24 +1,25 @@
-import { useState } from "react";
 import DishComponent from "./DishComponent";
-import { GetDishes } from "../services/DbService";
 import styled from "styled-components";
-
+import { useState } from "react";
+import { Dish } from "../Models/Dish";
+import { PostQuery } from "../services/DbService";
 interface dishInput {
   dishType: string;
 }
-
 export const DishListComponent = ({ dishType }: dishInput) => {
   const [selectedDish, setSelectedDish] = useState<number | null>(null);
-  const mainDish = GetDishes(dishType);
-  const isSideDish = dishType.toLowerCase() === "sidedish" ? true : false;
-
+  const isSideDish = dishType === "sideDish" ? true : false;
   const HandleClick = (index: number) => {
     setSelectedDish(index === selectedDish ? null : index);
   };
+  const { data, isLoading, error } = PostQuery(dishType);
+
+  if (isLoading) return <div>Loading...</div>;
+  if (error) return <div>Error: {error.message}</div>;
 
   return (
     <DishesContainer>
-      {mainDish?.map((dish, index) => (
+      {data?.map((dish: Dish, index: number) => (
         <DishComponent
           key={index}
           dish={dish}
